itest: revive test testSendMultiPathPayment
