+++ lntest/itest/lnd_mpp_test.go
@@ -1,6 +1,7 @@
 package itest
 
 import (
+	"encoding/hex"
 	"time"
 
 	"github.com/btcsuite/btcd/btcutil"
@@ -154,6 +155,101 @@ func testSendToRouteMultiPath(ht *lntest.HarnessTest) {
 
 	// ...and in Bob's list of paid invoices.
 	assertSettledInvoice(ctx.bob, rHash, 3)
+}
+
+// testSendMultiPathPayment tests that we are able to successfully route a
+// payment using multiple shards across different paths.
+func testSendMultiPathPayment(ht *lntest.HarnessTest) {
+	ctx := newMppTestContext(ht)
+	defer ctx.shutdownNodes()
+
+	const paymentAmt = btcutil.Amount(300000)
+
+	// Set up a network with three different paths Alice <-> Bob. Channel
+	// capacities are set such that the payment can only succeed if (at
+	// least) three paths are used.
+	//
+	//              _ Eve _
+	//             /       \
+	// Alice -- Carol ---- Bob
+	//      \              /
+	//       \__ Dave ____/
+	//
+	ctx.openChannel(ctx.carol, ctx.bob, 135000)
+	ctx.openChannel(ctx.alice, ctx.carol, 235000)
+	ctx.openChannel(ctx.dave, ctx.bob, 135000)
+	ctx.openChannel(ctx.alice, ctx.dave, 135000)
+	ctx.openChannel(ctx.eve, ctx.bob, 135000)
+	ctx.openChannel(ctx.carol, ctx.eve, 135000)
+
+	defer ctx.closeChannels()
+
+	// Increase Dave's fee to make the test deterministic. Otherwise it
+	// would be unpredictable whether pathfinding would go through Charlie
+	// or Dave for the first shard.
+	req := &lnrpc.PolicyUpdateRequest{
+		Scope:         &lnrpc.PolicyUpdateRequest_Global{Global: true},
+		BaseFeeMsat:   500000,
+		FeeRate:       0.001,
+		TimeLockDelta: 40,
+	}
+	ht.UpdateChannelPolicy(ctx.dave, req)
+
+	// Our first test will be Alice paying Bob using a SendPayment call.
+	// Let Bob create an invoice for Alice to pay.
+	payReqs, rHashes, invoices := ht.CreatePayReqs(ctx.bob, paymentAmt, 1)
+
+	rHash := rHashes[0]
+	payReq := payReqs[0]
+
+	sendReq := &routerrpc.SendPaymentRequest{
+		PaymentRequest: payReq,
+		MaxParts:       10,
+		TimeoutSeconds: 60,
+		FeeLimitMsat:   noFeeLimitMsat,
+	}
+	payment := ht.SendPaymentAndAssert(ctx.alice, sendReq)
+
+	// Make sure we got the preimage.
+	require.Equal(ht, hex.EncodeToString(invoices[0].RPreimage),
+		payment.PaymentPreimage, "preimage doesn't match")
+
+	// Check that Alice split the payment in at least three shards. Because
+	// the hand-off of the htlc to the link is asynchronous (via a mailbox),
+	// there is some non-determinism in the process. Depending on whether
+	// the new pathfinding round is started before or after the htlc is
+	// locked into the channel, different sharding may occur. Therefore we
+	// can only check if the number of shards isn't below the theoretical
+	// minimum.
+	succeeded := 0
+	for _, htlc := range payment.Htlcs {
+		if htlc.Status == lnrpc.HTLCAttempt_SUCCEEDED {
+			succeeded++
+		}
+	}
+
+	const minExpectedShards = 3
+	require.GreaterOrEqual(ht, succeeded, minExpectedShards,
+		"expected shards not reached")
+
+	// Make sure Bob show the invoice as settled for the full amount.
+	inv := ht.LookupInvoice(ctx.bob, rHash)
+
+	require.EqualValues(ht, paymentAmt, inv.AmtPaidSat,
+		"incorrect payment amt")
+
+	require.Equal(ht, lnrpc.Invoice_SETTLED, inv.State,
+		"Invoice not settled")
+
+	settled := 0
+	for _, htlc := range inv.Htlcs {
+		if htlc.State == lnrpc.InvoiceHTLCState_SETTLED {
+			settled++
+		}
+
+	}
+	require.Equal(ht, succeeded, settled,
+		"num of HTLCs wrong")
 }
 
 type mppTestContext struct {
